Working user authenticated jwt validation

+++ Backend/internal/store/models/product.go
@@ -6,7 +6,7 @@ type Product struct {
 	Description *string  `json:"description,omitempty"`
 	Price       float64  `json:"price"`
 	RentalPrice *float64 `json:"rental_price,omitempty"`
-	Color       uint32   `json:"color"`
+	Color       uint64   `json:"color"`
 	Size        float64  `json:"size"`
 	ImageURL    *string  `json:"image_url,omitempty"`
 	Stock       int      `json:"stock"`
